Mac port: PVR format supported, and 64-bit fixes
Using NSUinteger and NSInteger instead of int, since it is a 64-bit
"word" on Snow Leopard


Former-commit-id: 4e35e87eaacd5a7e226818a78e9dbebe188f01f0

+++ cocos2d/CCDirector.h
@@ -88,7 +88,7 @@ and when to execute the Scenes.
 	/* display FPS ? */
 	BOOL displayFPS_;
 
-	int frames_;
+	NSUInteger frames_;
 	ccTime accumDt_;
 	ccTime frameRate_;
 #if	CC_DIRECTOR_FAST_FPS
